mqtt_redis_app/requirements.txt
paho-mqtt==1.6.1
influxdb-client
